requirements.txt
chromadb>=0.6.3
httpx>=0.28.1
ipykernel>=6.30.0
langchain>=0.3.27
langchain-chroma>=0.2.6
langchain-community>=0.3.9
langchain-core>=0.3.72
langchain-openai>=0.3.28
langgraph>=0.5.4
matplotlib>=3.9.0
pandas>=2.2.3
python-dotenv>=1.1.1
sqlalchemy>=2.0.41
